scripts/requirements.txt
anthropic>=0.45.0
elevenlabs>=1.0.0
fal-client>=0.5.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
boto3>=1.34.0
requests>=2.31.0
Pillow>=10.0.0
